feat(key-result): decoupled user list
