+++ src/components/KeyResult/Single/Sections/Owner/queries.gql
